refactor filter tests; WIP on DeprecationWarnings; new filter_parent_fault_ids; detox
