+++ test/test_rupture_representatives.py
@@ -1,12 +1,4 @@
-import pytest
-
-from solvis.fault_system_solution_helper import FaultSystemSolutionHelper, build_rupture_groups
-
-
-@pytest.fixture
-def fss_helper(composite_fixture):
-    fss = composite_fixture._solutions['CRU']
-    yield FaultSystemSolutionHelper(fss)
+from solvis.fault_system_solution_helper import build_rupture_groups
 
 
 def test_build_rupture_groups(composite_fixture):
